requirements.txt
requests
browser-cookie3
pathvalidate
selenium
openpyxl
coloredlogs
portalocker
systemd-python
selenium-wire

